refactor(funds): lift fund/goal fetch to FundLibraryClient via useFundsData (#10)
Both views are always mounted (FundLibraryClient renders both, CSS-toggled), so
each one independently fetched /api/funds + /api/v1/savings-goals on mount (~4
requests) and reloaded on every mutation, kept in sync by a cross-view
'cairn:funds-updated' event.

New useFundsData hook owns funds + goals + loading + error + reload (and the
2-min cache). FundLibraryClient calls it once and spreads it into both views as
props. Because the two views now read one shared copy:
- the page fetches funds + goals a single time,
- a mutation in either view updates the shared state (and the hidden view
  reflects it without a reload), so the 'cairn:funds-updated' event and its
  _suppressNotify guard are deleted,
- the duplicated Fund/Goal/FundType types and cache helpers collapse into the
  hook (one definition instead of two).

Pure refactor — no behavior change. The existing funds E2E suite is the
regression gate: mobile funds (21), desktop funds incl. the cross-view
DCA-reflects-on-mobile test (20), input-zoom (4) all pass; plus 809 unit tests,
lint at the 53 baseline (0 new), tsc clean.
